requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
mysql-connector==2.2.9
requests==2.22.0
urllib3==1.25.3
